+++ conman/routes/admin.py
@@ -7,8 +7,6 @@ should really help in understanding the code here:
 https://django-polymorphic.readthedocs.io/en/stable/admin.html
 """
 
-from django.apps import apps
-from django.conf import settings
 from django.contrib import admin
 from polymorphic.admin import (
     PolymorphicChildModelAdmin,
@@ -48,6 +46,5 @@ class RouteParentAdmin(PolymorphicParentModelAdmin):
     search_fields = ('url',)
 
     def get_child_models(self):
-        """Return every admin-registered subclass of Route."""
-        paths = settings.CONMAN_ADMIN_ROUTES
-        return [apps.get_model(model_path) for model_path in paths]
+        """Return every subclass of Route."""
+        return list(self.model.get_subclasses())