+++ example/example/settings.py
@@ -63,10 +63,3 @@ TEMPLATES = [{
         ],
     },
 }]
-
-
-CONMAN_ADMIN_ROUTES = [
-    'example.Page',
-    'redirects.RouteRedirect',
-    'redirects.URLRedirect',
-]